+++ app/component-data/NavigationBar/joinFlowTabData.tsx
@@ -2,7 +2,7 @@ import EcosystemIcon from "~/ui/design-system/images/content/ecosystem"
 import AmbassadorsIcon from "~/ui/design-system/images/nav/ambassadors"
 import SocialsIcon from "~/ui/design-system/images/nav/socials"
 import { Section } from "~/ui/design-system/src/lib/Components/NavigationBar/types"
-import { DISCORD_URL } from "~/ui/design-system/src/lib/constants"
+import { DISCORD_URL } from "~/constants"
 import { emeraldDaoCard, flowVerseCard, grantsCard } from "./navCardsData"
 // Connect Tab
 export const eventsSection: Section = {
