+++ app/ui/design-system/src/lib/Pages/CommunityPage/index.tsx
@@ -16,7 +16,7 @@ import { LandingHeader } from "../../Components/LandingHeader"
 import ProjectCards, { ProjectCardsProps } from "../../Components/ProjectCards"
 import { ToolsAndConceptsProps } from "../../Components/ToolsAndConcepts"
 import { UpcomingEventsProps } from "../../Components/UpcomingEvents"
-import { DISCOURSE_URL, GITHUB_URL } from "../../constants"
+import { DISCOURSE_URL, GITHUB_URL } from "../../../../../../constants"
 import PageBackground from "../shared/PageBackground"
 import PageSection from "../shared/PageSection"
 import PageSections from "../shared/PageSections"
